First Commit
Practicing how to use search bar and the data is from online. I just used .filter.map.toLowerCase.includes.length it is all core practice in javascript. will practice to use it in actually project. planning to use a data from api and will implement it in the next commit